Server/requirements.txt
fastapi[standard]
aiokafka
cassandra-driver
